guidance: rename guidance_playground → guidance_gui, tidy repo layout
- Rename guidance_playground/ → guidance_gui/ (all py files, prototypes .npy,
  scripts/) and update all import strings and path literals accordingly
- Add guidance_gui/__init__.py and __main__.py so `python3 -m guidance_gui` works
- Use __file__-relative absolute default for prototypes path in build_guidance_panel()
  so the gallery loads correctly regardless of working directory
- Move DESIGN.md and GUIDANCE_FRAMEWORK_DESIGN.md to docs/ (renamed to
  guidance_gui_design.md / guidance_framework_design.md)
- Delete outdated documentation_guidance.md (pre-framework API tutorial)
- Remove diagnostic prototypes_k16_counts.npy (never loaded by application code)
- Add guidance_gui/README.md with launch instructions and prototypes note
- Add Design Rationale section to docs/guidance_gui_design.md explaining why
  the DPO GUI fixes noise_scale≈0 for the guided trajectory
