+++ guidance_gui/__main__.py
@@ -1,0 +1,3 @@
+from guidance_gui.app import main
+
+main()
